requirements.txt
anthropic>=0.7.0
pdf2image>=1.16.3
pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
PyQt6>=6.5.0